requirements.txt
# Advanced RAG Project - Core Dependencies

# API and AI
mistralai>=1.8.0

# PDF Processing
PyMuPDF>=1.26.0
PyPDF2>=3.0.0

# Vector Search and Embeddings
faiss-cpu>=1.11.0
numpy>=1.24.0

# NLP and Tokenization
spacy>=3.8.0
tiktoken>=0.9.0

# Configuration and Validation
pydantic>=2.0.0
python-dotenv>=1.0.0

# Web Interface
streamlit>=1.47.0
